bank: deposit and withdraw functionality
Added deposit and withdraw functionality, so that users can operate with their money . Username is hardcoded because there's still no session to get it from.

Fixes #75
Fixes #71

+++ src/main/java/com/clouway/core/Account.java
@@ -6,9 +6,9 @@ package com.clouway.core;
 public class Account {
   public final String name;
   public final String password;
-  public final Integer amount;
+  public final Double amount;
 
-  public Account(String name, String password, Integer amount) {
+  public Account(String name, String password, Double amount) {
     this.name = name;
     this.password = password;
     this.amount = amount;
